requirements.txt
streamlit==1.51.0
torch==2.6.0
torchvision==0.21.0
transformers==4.42.0
tokenizers==0.19.1
scikit-learn==1.7.2
pandas==2.3.3
numpy==1.26.4
pillow==11.3.0
joblib==1.5.2
huggingface-hub==0.36.0
accelerate==1.11.0